chore: splitted validations and schema
